requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
einops==0.8.1
eval_type_backport==0.2.2
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.7.0
future==1.0.0
grpcio==1.74.0
huggingface-hub==0.34.3
idna==3.10
imagecodecs==2024.12.30
imageio==2.37.0
importlib_metadata==8.7.0
Jinja2==3.1.6
kornia==0.6.4
lazy_loader==0.4
lightning-utilities==0.15.0
Markdown==3.8.2
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.6.3
networkx==3.2.1
numpy>=2.0,<2.3
omegaconf==2.3.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.1
pillow==11.3.0
propcache==0.3.2
protobuf==6.31.1
pydantic==2.11.7
pydantic_core==2.33.2
pyDeprecate==0.3.1
python-dateutil==2.9.0.post0
pytorch-lightning==1.4.2
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
scikit-image==0.24.0
scipy==1.13.1
simsimd==6.5.0
six==1.17.0
stringzilla==3.12.5
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
test_tube==0.7.5
tifffile==2024.8.30
tokenizers==0.12.1
torch==2.7.1
torchmetrics==0.6.0
torchvision==0.22.1
tqdm==4.67.1
transformers==4.19.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1
zipp==3.23.0
